Politicas aplicadas

+++ app/Http/Controllers/PostsController.php
@@ -53,6 +53,7 @@ class PostsController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('update', $post);
         return view('posts.edit')->with([
             'post' => $post
         ]);
@@ -63,6 +64,7 @@ class PostsController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->authorize('update', $post);
         $post->update($request->only('title', 'excerpt', 'content'));
 
         return redirect()->route('inicio');
@@ -73,6 +75,7 @@ class PostsController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
         $post->delete();
 
         return redirect()->route('inicio');
